# Reduce fileclient latency by merging _file_hash and _find_file