### What does this PR do?

1. Creates new master function _file_hash_and_stat
(sadly that's backwards incompatible, which means you need to update your master before you update minions)
2. Changes fileclient.py to use one call to _file_hash_and_stat instead of calling _file_hash and then _find_file.

This reduces latency to access file's metadata in minion<-> master setup when they are in different datacenter - in my case the ping latency is 10ms.

This somehow added to closer to 40ms than 20ms when requesting file metadata from master.
This commit reduced our deploy times - 7 machines - from 27 seconds to 22 seconds, by cutting access times in half.
(both of these timings include my not-yet-finished ZeroMQ connection caching patch from #39338, which saves additional 50x50ms = 2500ms, around 50 re-connections per highstate with 50ms handshake time, as well as a the new WIP plugin loader)

There is still big room for improvement in near future - I'd like to add last modified dates to _file_list, update file list once when a job is requested, and then latency to check if file is up to date would be 0ms. 🙃 

Note: file_hash_and_stat was an already existing function in filesystem module, but wasn't hooked to the public API for some reason 😕 

### What issues does this PR fix or reference?

Related to #39338, though this is a seperate problem.

### Previous Behavior
For any requested file that is up-to-date in cache, _file_hash would be requested and then _file_find.

### New Behavior
For any requested file that is up-to-date in cache, only one API request _file_hash_and_stat would be send instead of two.

### Tests written?

No

Please review [Salt's Contributing Guide](https://docs.saltstack.com/en/latest/topics/development/contributing.html) for best practices.