@terminalmage 2017-02-13
No, please don't merge this. It was written that way for a reason. We can't just stat the file as it exists locally on the master, because in the case of gitfs (the only fileserver backend other than roots which supports mode preservation in Salt) we need to get the mode of the file from the git repository's metadata for that file. So, we need to query the fileserver backend, which is why the RemoteClient needs to run the ``_file_find`` master func to have it query the fileserver to return the file mode.