@ninja- 2017-02-13
@terminalmage it supports file chmods correctly, I doubt this could break gitfs backend