@ninja- 2017-02-13
> ) we need to get the mode of the file from the git repository's metadata for that file