@ninja- 2017-02-13
It does exactly the same work look i.e. here https://github.com/saltstack/salt/blob/develop/salt/fileserver/__init__.py#L640