@ninja- 2017-02-13
If you are referring to https://github.com/ninja-/salt/blob/94c62388e792884cebc11095db20e3db81fa1348/salt/fileclient.py#L1278 that's only used when the url is local and without salt:// prefix