@ninja- 2017-02-13
> We can't just stat the file as it exists locally on the master, because in the case of gitfs (the only fileserver backend other than roots which supports mode preservation in Salt) we need to get the mode of the file from the git repository's metadata for that file. 

If that's really the case then the chmod should be preserved in master's cached file....