@ninja- 2017-02-13
@terminalmage I double checked and I guarantee you 100% that it's calling the same find_file function interally and returning it's stats from find_file(...)['stats'])