@terminalmage 2017-02-13
Yes, after looking at the full file, I can see that the ``os.stat()`` was being done for local files only. The stat logic is only being performed at this time for the mode preservation feature added in 2016.11.0 for file states.

Sorry for the misunderstanding, I misinterpreted the diff and thought you were getting the stats from the cached copy rather than the remote copy.