@twangboy 2017-03-08
@ninja- This PR breaks the state system on Windows. Could you take a look at it?
https://github.com/saltstack/salt/issues/39905